requirements.txt
# Core Framework
streamlit>=1.32.0
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
httpx>=0.27.0

# Audio Processing
librosa>=0.10.1
soundfile>=0.12.1
audio-recorder-streamlit>=0.0.8
numpy>=1.26.4
pandas>=2.2.0

# Machine Learning
torch>=2.2.0
torchaudio>=2.2.0
transformers>=4.39.0
accelerate>=0.28.0
openai-whisper>=20231117

# Database
sqlalchemy>=2.0.28

# Utilities
python-dotenv>=1.0.1
pydantic>=2.6.4
pydantic-settings>=2.2.1
requests>=2.31.0
anyio>=4.3.0

# HuggingFace Hub
huggingface_hub>=0.22.2
datasets>=2.18.0
